+++ packages/smarthr-ui/src/components/Checkbox/index.ts
@@ -1,1 +1,2 @@
 export { Checkbox, type Props } from './Checkbox'
+export { UnlabeledCheckbox } from './UnlabeledCheckbox'
