+++ packages/smarthr-ui/src/components/RadioButton/index.ts
@@ -1,1 +1,2 @@
 export { RadioButton } from './RadioButton'
+export { UnlabeledRadioButton } from './UnlabeledRadioButton'
